+++ R/zzz.R
@@ -24,7 +24,7 @@ R62Fun(DistributionWrapper,
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage("\n-----------------------------")
   packageStartupMessage(
-    "\tdistr6 v 1.4.2",
+    "\tdistr6 v 1.4.3",
     "\n\nGet started:\t?distr6
 Changelog:\tdistr6News()"
   )